bl-interop-client/src/app/SubscriberTopic/SubscriberTopic.component.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Component, OnInit, Input } from '@angular/core';
import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
import { SubscriberTopicService } from './SubscriberTopic.service';
import 'rxjs/add/operator/toPromise';

@Component({
  selector: 'app-subscribertopic',
  templateUrl: './SubscriberTopic.component.html',
  styleUrls: ['./SubscriberTopic.component.css'],
  providers: [SubscriberTopicService]
})
export class SubscriberTopicComponent implements OnInit {

  myForm: FormGroup;

  private allAssets;
  private asset;
  private currentId;
  private errorMessage;

  topicSubscriberId = new FormControl('', Validators.required);
  topic = new FormControl('', Validators.required);
  subscriber = new FormControl('', Validators.required);

  constructor(public serviceSubscriberTopic: SubscriberTopicService, fb: FormBuilder) {
    this.myForm = fb.group({
      topicSubscriberId: this.topicSubscriberId,
      topic: this.topic,
      subscriber: this.subscriber
    });
  };

  ngOnInit(): void {
    this.loadAll();
  }

  loadAll(): Promise<any> {
    const tempList = [];
    return this.serviceSubscriberTopic.getAll()
    .toPromise()
    .then((result) => {
      this.errorMessage = null;
      result.forEach(asset => {
        tempList.push(asset);
      });
      this.allAssets = tempList;
    })
    .catch((error) => {
      if (error === 'Server error') {
        this.errorMessage = 'Could not connect to REST server. Please check your configuration details';
      } else if (error === '404 - Not Found') {
        this.errorMessage = '404 - Could not find API route. Please check your available APIs.';
      } else {
        this.errorMessage = error;
      }
    });
  }

	/**
   * Event handler for changing the checked state of a checkbox (handles array enumeration values)
   * @param {String} name - the name of the asset field to update
   * @param {any} value - the enumeration value for which to toggle the checked state
   */
  changeArrayValue(name: string, value: any): void {
    const index = this[name].value.indexOf(value);
    if (index === -1) {
      this[name].value.push(value);
    } else {
      this[name].value.splice(index, 1);
    }
  }

	/**
	 * Checkbox helper, determining whether an enumeration value should be selected or not (for array enumeration values
   * only). This is used for checkboxes in the asset updateDialog.
   * @param {String} name - the name of the asset field to check
   * @param {any} value - the enumeration value to check for
   * @return {Boolean} whether the specified asset field contains the provided value
   */
  hasArrayValue(name: string, value: any): boolean {
    return this[name].value.indexOf(value) !== -1;
  }

  addAsset(form: any): Promise<any> {
    this.asset = {
      $class: 'org.bl.network.SubscriberTopic',
      'topicSubscriberId': this.topicSubscriberId.value,
      'topic': this.topic.value,
      'subscriber': this.subscriber.value
    };

    this.myForm.setValue({
      'topicSubscriberId': null,
      'topic': null,
      'subscriber': null
    });

    return this.serviceSubscriberTopic.addAsset(this.asset)
    .toPromise()
    .then(() => {
      this.errorMessage = null;
      this.myForm.setValue({
        'topicSubscriberId': null,
        'topic': null,
        'subscriber': null
      });
      this.loadAll();
    })
    .catch((error) => {
      if (error === 'Server error') {
          this.errorMessage = 'Could not connect to REST server. Please check your configuration details';
      } else {
          this.errorMessage = error;
      }
    });
  }


  updateAsset(form: any): Promise<any> {
    this.asset = {
      $class: 'org.bl.network.SubscriberTopic',
      'topic': this.topic.value,
      'subscriber': this.subscriber.value
    };

    return this.serviceSubscriberTopic.updateAsset(form.get('topicSubscriberId').value, this.asset)
    .toPromise()
    .then(() => {
      this.errorMessage = null;
      this.loadAll();
    })
    .catch((error) => {
      if (error === 'Server error') {
        this.errorMessage = 'Could not connect to REST server. Please check your configuration details';
      } else if (error === '404 - Not Found') {
        this.errorMessage = '404 - Could not find API route. Please check your available APIs.';
      } else {
        this.errorMessage = error;
      }
    });
  }


  deleteAsset(): Promise<any> {

    return this.serviceSubscriberTopic.deleteAsset(this.currentId)
    .toPromise()
    .then(() => {
      this.errorMessage = null;
      this.loadAll();
    })
    .catch((error) => {
      if (error === 'Server error') {
        this.errorMessage = 'Could not connect to REST server. Please check your configuration details';
      } else if (error === '404 - Not Found') {
        this.errorMessage = '404 - Could not find API route. Please check your available APIs.';
      } else {
        this.errorMessage = error;
      }
    });
  }

  setId(id: any): void {
    this.currentId = id;
  }

  getForm(id: any): Promise<any> {

    return this.serviceSubscriberTopic.getAsset(id)
    .toPromise()
    .then((result) => {
      this.errorMessage = null;
      const formObject = {
        'topicSubscriberId': null,
        'topic': null,
        'subscriber': null
      };

      if (result.topicSubscriberId) {
        formObject.topicSubscriberId = result.topicSubscriberId;
      } else {
        formObject.topicSubscriberId = null;
      }

      if (result.topic) {
        formObject.topic = result.topic;
      } else {
        formObject.topic = null;
      }

      if (result.subscriber) {
        formObject.subscriber = result.subscriber;
      } else {
        formObject.subscriber = null;
      }

      this.myForm.setValue(formObject);

    })
    .catch((error) => {
      if (error === 'Server error') {
        this.errorMessage = 'Could not connect to REST server. Please check your configuration details';
      } else if (error === '404 - Not Found') {
        this.errorMessage = '404 - Could not find API route. Please check your available APIs.';
      } else {
        this.errorMessage = error;
      }
    });
  }

  resetForm(): void {
    this.myForm.setValue({
      'topicSubscriberId': null,
      'topic': null,
      'subscriber': null
      });
  }

}
